Added fixture pxshell and keepass_db
